requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
dm-tree==0.1.8
Farama-Notifications==0.0.4
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.4.0
GPUtil==1.4.0
gymnasium==0.29.1
idna==3.6
importlib_metadata==7.0.2
importlib_resources==6.3.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.3
mpmath==1.3.0
msgpack==1.0.8
networkx==3.0
numpy==1.24.4
packaging==24.0
pillow==10.2.0
pkgutil_resolve_name==1.3.10
protobuf==5.26.0
python-dotenv==1.0.1
PyYAML==6.0
ray[data,train,tune,rllib]==2.24.0
referencing==0.34.0
rel==0.4.8
requests==2.31.0
rpds-py==0.18.0
sympy==1.12
torch==2.2.1
typing_extensions==4.10.0
urllib3==2.2.1
websocket-client==1.5.1
zipp==3.18.1
